+++ crates/matrix-sdk/src/room/mod.rs
@@ -26,7 +26,9 @@ use std::{
 use async_stream::stream;
 use eyeball::SharedObservable;
 use futures_core::Stream;
-use futures_util::{future::join_all, stream::FuturesUnordered};
+use futures_util::{
+    future::join_all, stream as futures_stream, stream::FuturesUnordered, StreamExt,
+};
 use http::StatusCode;
 #[cfg(feature = "e2e-encryption")]
 pub use identity_status_changes::IdentityStatusChanges;
@@ -127,7 +129,6 @@ use ruma::{
 use serde::de::DeserializeOwned;
 use thiserror::Error;
 use tokio::{join, sync::broadcast};
-use tokio_stream::StreamExt;
 use tracing::{debug, error, info, instrument, trace, warn};
 
 use self::futures::{SendAttachment, SendMessageLikeEvent, SendRawMessageLikeEvent};
@@ -320,13 +321,16 @@ impl Room {
     /// Only invited and joined rooms can be left.
     #[doc(alias = "reject_invitation")]
     #[instrument(skip_all, fields(room_id = ?self.inner.room_id()))]
-    pub async fn leave(&self) -> Result<()> {
+    async fn leave_impl(&self) -> (Result<()>, &Room) {
         let state = self.state();
         if state == RoomState::Left {
-            return Err(Error::WrongRoomState(Box::new(WrongRoomState::new(
-                "Joined or Invited",
-                state,
-            ))));
+            return (
+                Err(Error::WrongRoomState(Box::new(WrongRoomState::new(
+                    "Joined or Invited",
+                    state,
+                )))),
+                self,
+            );
         }
 
         // If the room was in Invited state we should also forget it when declining the
@@ -354,11 +358,13 @@ impl Room {
             error!(?error, ignore_error, should_forget, "Failed to leave the room");
 
             if !ignore_error {
-                return Err(error.into());
+                return (Err(error.into()), self);
             }
         }
 
-        self.client.base_client().room_left(self.room_id()).await?;
+        if let Err(e) = self.client.base_client().room_left(self.room_id()).await {
+            return (Err(e.into()), self);
+        }
 
         if should_forget {
             trace!("Trying to forget the room");
@@ -368,7 +374,59 @@ impl Room {
             }
         }
 
-        Ok(())
+        (Ok(()), self)
+    }
+
+    /// Leave this room and all predecessors.
+    /// If any room was in [`RoomState::Invited`] state, it'll also be forgotten
+    /// automatically.
+    ///
+    /// Only invited and joined rooms can be left.
+    /// Will return an error if the current room fails to leave but
+    /// will only warn if a predecessor fails to leave.
+    pub async fn leave(&self) -> Result<()> {
+        let mut rooms: Vec<Room> = vec![self.clone()];
+        let mut current_room = self;
+
+        while let Some(predecessor) = current_room.predecessor_room() {
+            let maybe_predecessor_room = current_room.client.get_room(&predecessor.room_id);
+
+            if let Some(predecessor_room) = maybe_predecessor_room {
+                rooms.push(predecessor_room.clone());
+                current_room = rooms.last().expect("Room just pushed so can't be empty");
+            } else {
+                warn!("Cannot find predecessor room");
+                break;
+            }
+        }
+
+        let batch_size = 5;
+
+        let rooms_futures: Vec<_> = rooms
+            .iter()
+            .filter_map(|room| match room.state() {
+                RoomState::Joined | RoomState::Invited | RoomState::Knocked => {
+                    Some(room.leave_impl())
+                }
+                RoomState::Banned | RoomState::Left => None,
+            })
+            .collect();
+
+        let mut futures_stream = futures_stream::iter(rooms_futures).buffer_unordered(batch_size);
+
+        let mut maybe_this_room_failed_with: Option<Error> = None;
+
+        while let Some(result) = futures_stream.next().await {
+            if let (Err(e), room) = result {
+                if room.room_id() == self.room_id() {
+                    maybe_this_room_failed_with = Some(e);
+                } else {
+                    warn!("Failure while attempting to leave predecessor room: {e:?}");
+                }
+            }
+        }
+
+        maybe_this_room_failed_with.map_or(Ok(()), Err)
     }
 
     /// Join this room.